+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- parameter to configure docker stop timeout duration.
+
 ## [1.6.1]
 ### Added
 - support for instance not found errors in gcp, by [frebib](https://github.com/frebib).